split URLSessionProtocol into it's own file.
