+++ Relay/TestUtilities.swift
@@ -8,6 +8,7 @@
 
 import Foundation
 
+
 func isRunningUnitTests() -> Bool {
     return ProcessInfo.processInfo.environment["XCTestConfigurationFilePath"] != nil
 }
