+++ RelayTests/URLSessionMock.swift
@@ -1,17 +1,15 @@
-///---
+//
+//  URLSessionProtocol.swift
+//  Relay
+//
+//  Created by Evan Kimia on 5/2/17.
+//  Copyright © 2017 zero. All rights reserved.
+//
 
 import Foundation
 
-public protocol URLSessionProtocol {
+@testable import Relay
 
-    var delegate: URLSessionDelegate? { get }
-
-    func uploadTask(with request: URLRequest, fromFile fileURL: URL) -> URLSessionUploadTask
-
-    func getAllTasks(completionHandler: @escaping ([URLSessionTask]) -> Swift.Void)
-}
-
-extension URLSession: URLSessionProtocol { }
 
 public final class URLSessionMock : URLSessionProtocol {
